kernel: lib: floating proportions
Upstream 145ca25e

Given a set of objects, floating proportions aims to efficiently give the
proportional 'activity' of a single item as compared to the whole set. Where
'activity' is a measure of a temporal property of the items.

It is efficient in that it need not inspect any other items of the set
in order to provide the answer. It is not even needed to know how many
other items there are.

It has one parameter, and that is the period of 'time' over which the
'activity' is measured.

170-lib-proportions.patch commits:
 afd7921 [RT-N] kernel: mm: BDI updates from upstream
 325d6d6 kernel: Add percpu_counter_sub() inline