kernel: vfs: Add 64 bit i_version support
Upstream 7a224228

The i_version field of the inode is changed to be a 64-bit counter that
is set on every inode creation and that is incremented every time the
inode data is modified (similarly to the "ctime" time-stamp).
The aim is to fulfill a NFSv4 requirement for rfc3530.
This first part concerns the vfs, it converts the 32-bit i_version in
the generic inode to a 64-bit, a flag is added in the super block in
order to check if the feature is enabled and the i_version is
incremented in the vfs.

144-vfs1-iversion64.patch commits:
 9120928 kernel: vfs: Add 64 bit i_version support
 75cf35f kernel: vfs: i_version: remount support
 75d24fb kernel: vfs: Fix sys_sync() and fsync_super() reliability (version 4)