kernel: is_vmalloc_addr(): Check if an address is within the vmalloc boundaries
Upstream 9e2779fa

Checking if an address is a vmalloc address is done in a couple of places.
Define a common version in mm.h and replace the other checks.

Again the include structures suck.  The definition of VMALLOC_START and
VMALLOC_END is not available in vmalloc.h since highmem.c cannot be included
there.