kernel: maps4: add proportional set size accounting in smaps
Upstream ec4dd3eb

The "proportional set size" (PSS) of a process is the count of pages it has
in memory, where each page is divided by the number of processes sharing
it.  So if a process has 1000 pages all to itself, and 1000 shared with one
other process, its PSS will be 1500.

               - lwn.net: "ELC: How much memory are applications really using?"

The PSS proposed by Matt Mackall is a very nice metic for measuring an
process's memory footprint.  So collect and export it via
/proc/<pid>/smaps.

Matt Mackall's pagemap/kpagemap and John Berthels's exmap can also do the
job.  They are comprehensive tools.  But for PSS, let's do it in the simple
way.

157-fs-proc-maps4-pss.patch commits:
 112e094 kernel: maps4: add proportional set size accounting in smaps
 86c157f kernel: maps4: move is_swap_pte