kernel: Embed a struct path into struct nameidata instead of nd->{dentry,mnt}
[ Upstream 4ac9137858 ]

This is the central patch of a cleanup series. In most cases there is no good
reason why someone would want to use a dentry for itself. This series reflects
that fact and embeds a struct path into nameidata.

Together with the other patches of this series
- it enforced the correct order of getting/releasing the reference count on
  <dentry,vfsmount> pairs
- it prepares the VFS for stacking support since it is essential to have a
  struct path in every place where the stack can be traversed
- it reduces the overall code size

This patch:

Switch from nd->{dentry,mnt} to nd->path.{dentry,mnt} everywhere.