kernel: ipv6: Do not drop packet if skb->local_df is set to true
[ Upstream 283d07ac20 ]

The old code will drop IPv6 packet if ipfragok is not set, since
ipfragok is obsoleted, will be instead by used skb->local_df, so this
check must be changed to skb->local_df.

This patch fix this problem and not drop packet if skb->local_df is
set to true.