igmpproxy: restrict igmp reports for downstream interfaces (wrt #2833339)
atm all igmp membership reports are forwarded to the upstream interface.
Unfortunately some ISP Providers restrict some multicast groups (esp. those
that are defined as local link groups and that are not supposed to be
forwarded to the wan, i.e 224.0.0.0/24). Therefore there should be some
kind of black oder whitelisting.
As whitelisting can be accomplished quite easy I wrote a litte patch, which
is attached to this request.

020-whitelist.patch commits:
 1fdcc2f igmpproxy 0.1_beta5
 aeb8998 igmpproxy: Restrict igmp reports forwarding to upstream interface