kernel: MIPS: Preliminary VDSO
Upstream c52d0d30

This is a preliminary patch to add a vdso to all user processes.  Still
missing are ELF headers and .eh_frame information.  But it is enough to
allow us to move signal trampolines off of the stack.  Note that emulation
of branch delay slots in the FPU emulator still requires the stack.

We allocate a single page (the vdso) and write all possible signal
trampolines into it.  The stack is moved down by one page and the vdso is
mapped into this space.

Patchwork: http://patchwork.linux-mips.org/patch/975/
	http://patchwork.linux-mips.org/patch/1385/
	http://patchwork.linux-mips.org/patch/1386/

171-vdso1.patch commits:
 fcf0659 [RT-N] kernel: MIPS: Preliminary VDSO support
 3e94753 kernel: MIPS: uasm: Updates from upstream